Extract POS refund flow state into POSRefundController
POSOrderListController keeps order selection and refund details for display.
POSRefundController owns the selection, review preparation and submission for
the order being refunded. POSOrderListModel coordinates the two.
